+++ PetrolPumpERP/Models/DataEntities/tblPurchaseInvoiceDetail.cs
@@ -16,7 +16,7 @@ namespace PetrolPumpERP.Models.DataEntities
     {
         public int PINoSrNo { get; set; }
         public Nullable<long> ProductCode { get; set; }
-        public Nullable<int> InvoiceQty { get; set; }
+        public Nullable<decimal> InvoiceQty { get; set; }
         public Nullable<decimal> InvoicePrice { get; set; }
         public Nullable<System.DateTime> ExpiryDate { get; set; }
         public string BatchNo { get; set; }
